refactor(web): home toggleInSet in lib/set, not lib/utils
toggleInSet is a Set-membership helper; park it in a module named for
what it holds rather than the catch-all utils.ts. utils.ts reverts to
just the shadcn cn helper.
